feat: Shimaas styling, integration of ratings, some main service fixes

+++ services/main_service/api/source/blueprints/user_rating.py
@@ -1,7 +1,9 @@
 from ..graph_db import GraphDb
+import json
+import uuid
 from ..helpers import db_helpers, request_helpers
 from flask import (
-    Blueprint, flash, g, redirect, request, session, url_for
+    Blueprint, request
 )
 
 
@@ -18,7 +20,7 @@ def create_rating(db_conn, user_id, rate_user_id, game_id, attr_id, skill_id):
     rating_inserted = db_conn.run(
         """
         MATCH(user: User {id: $from_user_id }), (target: User {id: $to_user_id}), (game: Game {id: $game_id}), 
-        (attr: Attribute { id: $attribute_id}), (skill: Skill {id: $skill_id})
+        (attr: Attribute {id: $attribute_id}), (skill: Skill {id: $skill_id})
         CREATE(rating: Rating { id: $id, from_user_id: $from_user_id, to_user_id: $to_user_id, game_id: $game_id, attribute_id: $attribute_id, skill_id: $skill_id }),
         (user)-[:HAS_GIVEN_RATING]->(rating),(rating)-[:HAS_GOTTEN_RATING]->(target),
         (rating)-[:FOR_GAME]->(game),(rating)-[:HAS_ATTR]->(attr),(rating)-[:HAS_SKILL]->(skill)
@@ -29,7 +31,7 @@ def create_rating(db_conn, user_id, rate_user_id, game_id, attr_id, skill_id):
     return rating_inserted
 
 
-@bp.get('/add')
+@bp.post('/add')
 def add_user_rating():
 
     if 'rate_user_id' not in request.form:
@@ -115,13 +117,8 @@ def list_users_followed():
 def delete_rating(db_conn, rating_id):
     attr_deleted = db_conn.run(
         """
-        MATCH
-        (rating: Rating{id: $id})-[given_rel:HAS_GIVEN_RATING]->(:User),
-        (rating)-[gotten_rel:HAS_GOTTEN_RATING]->(:User),
-        (rating)-[game_rel: FOR_GAME]->(:Game), 
-        (rating)-[attr_rel: HAS_ATTR]->(:Attribute), 
-        (rating)-[skill_rel:WITH_SKILL]->(:Skill)
-        DELETE given_rel, gotten_rel, game_rel, attr_rel, skill_rel, rating
+        MATCH (rating: Rating {id: $id})
+        DETACH DELETE rating
         RETURN rating
         """, {'id': rating_id}
     ).data()
@@ -142,36 +139,6 @@ def list_ratings():
     ).data()
 
     return ratings_query
-
-
-@bp.post('/add')
-def add_ratings():
-    if 'game_id' not in request.form:
-        return 'Error: Missing form field { game_id }'
-
-    if 'attribute_id' not in request.form:
-        return 'Error: Missing form field { attribute_id }'
-
-    if 'skill_id' not in request.form:
-        return 'Error: Missing form field { skill_id }'
-
-    graph = GraphDb()
-    db_conn = graph.get_database_driver()
-
-    game_id = request.form['game_id']
-    if not db_helpers.game_id_exists(db_conn, game_id):
-        return 'Error: A game with the given id does not exist'
-
-    attr_id = request.form['attribute_id']
-    if not db_helpers.attr_id_exists(db_conn, attr_id):
-        return 'Error: An attribute with the given id does not exist'
-
-    skill_id = request.form['skill_id']
-    if not db_helpers.skill_id_exists(db_conn, skill_id):
-        return 'Error: A skill with the given id does not exist'
-
-    return create_rating(db_conn, game_id, attr_id, skill_id)
-
 
 @bp.put('/edit')
 def edit_rating():
@@ -197,32 +164,16 @@ def edit_rating():
     if not db_helpers.skill_id_exists(db_conn, skill_id):
         return 'Error: A skill with the given id does not exist'
 
-    rating = db_conn.run(
-        """
-        MATCH(rating: Rating {id: $id})
-        RETURN rating
-        """, {'id': rating_id}
-    ).single()
+    user_id = request_helpers.get_user_id(request.headers)
 
-    user_id = rating['from_user_id']
-    rate_user_id = rating['to_user_id']
+    rate_user_id = request.form['rate_user_id']
+    if not db_helpers.user_id_exists(db_conn, rate_user_id):
+        return 'Error: A user with the given id does not exist'
 
     # Just delete the existing rating and recreate it. id don't matter.
     delete_rating(db_conn, rating_id)
     rating_created = create_rating(db_conn, user_id, rate_user_id,  game_id, attr_id, skill_id)
     return rating_created
-
-
-@bp.delete('/delete')
-def delete_rating():
-    if 'rating_id' not in request.form:
-        return 'Error: Missing form field { rating_id }'
-
-    graph = GraphDb()
-    db_conn = graph.get_database_driver()
-
-    return delete_rating(db_conn, request.form['rating_id'])
-
 
 class Rating:
     def __init__(self, from_user_id, to_user_id, game_id, attribute_id, skill_id):
